requirements.txt
APScheduler==2.1.2
Django==1.6.5
Jinja2==2.7.3
Markdown==2.4.1
MarkupSafe==0.23
Pygments==1.6
SQLAlchemy==0.9.6
Sphinx==1.2.3
defusedxml==0.4.1
django-admin-ext==1.0.2
django-filter==0.7
django-nose==1.2
djangorestframework==2.3.14
django-picklefield==0.3.1
docutils==0.12
dropbox==2.1.0
futures==2.1.6
nose==1.3.3
oauthlib==0.6.3
python-openid==2.2.5
python-social-auth==0.1.26
pytz==2014.4
requests==2.3.0
requests-oauthlib==0.4.1
six==1.7.3
tzlocal==1.1.1
urllib3==1.9